docs: Update guides, and references.txt
I also found this resource underTheHoodOfExecutables, and started looking
at the chapter - Why is my "Hello World" so Big?. I wrote the hello_world
in C and looked a bit into the compilation/assembling workflow with clang/LLVM
and the disassembling of the binary with otool and nm. With ChatGPT, a small
write up was completed on the clang, otool and nm commands for review and
updating as I learn more about this workflow. See hello_world folder in
programs.